requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
filelock==3.20.0
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==2.1.5
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
pillow==12.0.0
regex==2025.11.3
requests==2.32.5
setuptools==70.2.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.1+cu130
torchaudio==2.9.1+cu130
torchvision==0.24.1+cu130
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2